+++ next-docs/public/styledExamples/icons/Generic.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   GenericAbout,
   GenericAlarm,
@@ -51,6 +50,7 @@ import {
   GenericTrophy,
   GenericUpload,
   GenericUser,
+  GenericUserSwapping,
   GenericUsers,
 } from '@heathmont/moon-icons';
 import Icon from './Icon';
@@ -219,6 +219,9 @@ const Example = ({ wrapperProps, props }: Props) => (
     </Icon>
     <Icon {...wrapperProps} name="GenericUser">
       <GenericUser {...props} />
+    </Icon>
+    <Icon {...wrapperProps} name="GenericUserSwapping">
+      <GenericUserSwapping {...props} />
     </Icon>
     <Icon {...wrapperProps} name="GenericUsers">
       <GenericUsers {...props} />