+++ next-docs/public/styledExamples/icons/Travel.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   TravelAirplane,
   TravelAirplaneDown,
@@ -18,6 +17,7 @@ import {
   TravelSuitcase,
   TravelSwimmingPool,
   TravelWiFi,
+  TravelYacht,
 } from '@heathmont/moon-icons';
 import Icon from './Icon';
 
@@ -86,6 +86,9 @@ const Example = ({ wrapperProps, props }: Props) => (
     </Icon>
     <Icon {...wrapperProps} name="TravelWiFi">
       <TravelWiFi {...props} />
+    </Icon>
+    <Icon {...wrapperProps} name="TravelYacht">
+      <TravelYacht {...props} />
     </Icon>
   </>
 );