+++ packages/icons/src/icons/GenericPlus.tsx
@@ -12,8 +12,9 @@ const Svg = (props: React.SVGProps<SVGSVGElement>) => (
     {...props}
   >
     <path
-      d="M21.115 16.5a.5.5 0 000-1v1zm-10.23-1a.5.5 0 000 1v-1zm4.615 5.616a.5.5 0 001 0h-1zm1-10.231a.5.5 0 00-1 0h1zm4.615 4.615h-10.23v1h10.23v-1zM16.5 21.116V10.885h-1v10.23h1zM25 16a9 9 0 01-9 9v1c5.523 0 10-4.477 10-10h-1zm-9 9a9 9 0 01-9-9H6c0 5.523 4.477 10 10 10v-1zm-9-9a9 9 0 019-9V6C10.477 6 6 10.477 6 16h1zm9-9a9 9 0 019 9h1c0-5.523-4.477-10-10-10v1z"
-      fill="currentColor"
+      d="M21.115 16h-10.23M16 21.116V10.885M25.5 16a9.5 9.5 0 11-19 0 9.5 9.5 0 0119 0z"
+      stroke="currentColor"
+      strokeLinecap="round"
     />
   </svg>
 );