+++ next-docs/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Change Log
 
+## 9.24.1
+
+### Patch Changes
+
+- Update SC icons, fix SC Accordion
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.24.1
+  - @heathmont/moon-icons@9.24.1
+  - @heathmont/moon-icons-tw@9.24.1
+  - @heathmont/moon-accordion@9.24.1
+  - @heathmont/moon-assets@9.24.1
+  - @heathmont/moon-charts@9.24.1
+  - @heathmont/moon-components@9.24.1
+  - @heathmont/moon-datepicker@9.24.1
+  - @heathmont/moon-draggabletable@9.24.1
+  - @heathmont/moon-pagination@9.24.1
+  - @heathmont/moon-popover@9.24.1
+  - @heathmont/moon-select@9.24.1
+  - @heathmont/moon-sidebar@9.24.1
+  - @heathmont/moon-table@9.24.1
+  - @heathmont/moon-themes@9.24.1
+  - @heathmont/moon-utils@9.24.1
+  - @heathmont/moon-core-tw@9.24.1
+  - @heathmont/moon-table-tw@9.24.1
+  - @heathmont/moon-themes-tw@9.24.1
+
 ## 9.24.0
 
 ### Minor Changes